requirements.txt
channels==3.0.3
django-cors-headers==3.13.0
djangorestframework==3.13.0
django-model-utils==4.2.0
django-import-export==3.0.0
djangorestframework-jwt==1.11.0
Django==3.1.5
fcm-django==1.0.12
razorpay==1.3.0
sentry-sdk==1.12.0
